Put cmdline utils in order
