+++ lookout/style/typos/preprocessing/__main__.py
@@ -1,0 +1,7 @@
+"""Command line tools entry module."""
+import sys
+
+from lookout.style.typos.preprocessing.cmdline_utils import main
+
+if __name__ == "__main__":
+    sys.exit(main())
